+++ rat-event/src/crossterm.rs
@@ -2,9 +2,20 @@
 
 #![allow(dead_code)]
 
+pub mod event {
+    pub use ratatui_crossterm::crossterm::event::Event;
+    pub use ratatui_crossterm::crossterm::event::KeyCode;
+    pub use ratatui_crossterm::crossterm::event::KeyEvent;
+    pub use ratatui_crossterm::crossterm::event::KeyEventKind;
+    pub use ratatui_crossterm::crossterm::event::KeyModifiers;
+    pub use ratatui_crossterm::crossterm::event::MouseButton;
+    pub use ratatui_crossterm::crossterm::event::MouseEvent;
+    pub use ratatui_crossterm::crossterm::event::MouseEventKind;
+}
+
 /// A copy of the crossterm-KeyModifiers. Plus a few combinations of modifiers.
 pub mod modifiers {
-    use ratatui_crossterm::crossterm::event::KeyModifiers;
+    use crate::crossterm::event::KeyModifiers;
 
     pub const NONE: KeyModifiers = KeyModifiers::NONE;
     pub const CONTROL: KeyModifiers = KeyModifiers::CONTROL;
@@ -85,160 +96,160 @@ pub mod modifiers {
 #[macro_export]
 macro_rules! ct_event {
     (key press $keychar:pat) => {
-        ratatui_crossterm::crossterm::event::Event::Key(ratatui_crossterm::crossterm::event::KeyEvent {
-            code: ratatui_crossterm::crossterm::event::KeyCode::Char($keychar),
+        $crate::crossterm::event::Event::Key($crate::crossterm::event::KeyEvent {
+            code: $crate::crossterm::event::KeyCode::Char($keychar),
             modifiers: $crate::crossterm::modifiers::NONE,
-            kind: ratatui_crossterm::crossterm::event::KeyEventKind::Press|ratatui_crossterm::crossterm::event::KeyEventKind::Repeat,
+            kind: $crate::crossterm::event::KeyEventKind::Press|$crate::crossterm::event::KeyEventKind::Repeat,
             ..
         })
     };
     (key press ANY-$keychar:pat) => {
-        ratatui_crossterm::crossterm::event::Event::Key(ratatui_crossterm::crossterm::event::KeyEvent {
-            code: ratatui_crossterm::crossterm::event::KeyCode::Char($keychar),
-            kind: ratatui_crossterm::crossterm::event::KeyEventKind::Press|ratatui_crossterm::crossterm::event::KeyEventKind::Repeat,
+        $crate::crossterm::event::Event::Key($crate::crossterm::event::KeyEvent {
+            code: $crate::crossterm::event::KeyCode::Char($keychar),
+            kind: $crate::crossterm::event::KeyEventKind::Press|$crate::crossterm::event::KeyEventKind::Repeat,
             ..
         })
     };
     (key press $mod:ident-$keychar:pat) => {
-        ratatui_crossterm::crossterm::event::Event::Key(ratatui_crossterm::crossterm::event::KeyEvent {
-            code: ratatui_crossterm::crossterm::event::KeyCode::Char($keychar),
+        $crate::crossterm::event::Event::Key($crate::crossterm::event::KeyEvent {
+            code: $crate::crossterm::event::KeyCode::Char($keychar),
             modifiers: $crate::crossterm::modifiers::$mod,
-            kind: ratatui_crossterm::crossterm::event::KeyEventKind::Press|ratatui_crossterm::crossterm::event::KeyEventKind::Repeat,
+            kind: $crate::crossterm::event::KeyEventKind::Press|$crate::crossterm::event::KeyEventKind::Repeat,
             ..
         })
     };
     (key release $keychar:pat) => {
-        ratatui_crossterm::crossterm::event::Event::Key(ratatui_crossterm::crossterm::event::KeyEvent {
-            code: ratatui_crossterm::crossterm::event::KeyCode::Char($keychar),
+        $crate::crossterm::event::Event::Key($crate::crossterm::event::KeyEvent {
+            code: $crate::crossterm::event::KeyCode::Char($keychar),
             modifiers: $crate::crossterm::modifiers::NONE,
-            kind: ratatui_crossterm::crossterm::event::KeyEventKind::Release,
+            kind: $crate::crossterm::event::KeyEventKind::Release,
             ..
         })
     };
     (key release ANY-$keychar:pat) => {
-        ratatui_crossterm::crossterm::event::Event::Key(ratatui_crossterm::crossterm::event::KeyEvent {
-            code: ratatui_crossterm::crossterm::event::KeyCode::Char($keychar),
-            kind: ratatui_crossterm::crossterm::event::KeyEventKind::Release,
+        $crate::crossterm::event::Event::Key($crate::crossterm::event::KeyEvent {
+            code: $crate::crossterm::event::KeyCode::Char($keychar),
+            kind: $crate::crossterm::event::KeyEventKind::Release,
             ..
         })
     };
     (key release $mod:ident-$keychar:pat) => {
-        ratatui_crossterm::crossterm::event::Event::Key(ratatui_crossterm::crossterm::event::KeyEvent {
-            code: ratatui_crossterm::crossterm::event::KeyCode::Char($keychar),
+        $crate::crossterm::event::Event::Key($crate::crossterm::event::KeyEvent {
+            code: $crate::crossterm::event::KeyCode::Char($keychar),
             modifiers: $crate::crossterm::modifiers::$mod,
-            kind: ratatui_crossterm::crossterm::event::KeyEventKind::Release,
+            kind: $crate::crossterm::event::KeyEventKind::Release,
             ..
         })
     };
 
     (keycode press F($code:literal)) => {
-        ratatui_crossterm::crossterm::event::Event::Key(ratatui_crossterm::crossterm::event::KeyEvent {
-            code: ratatui_crossterm::crossterm::event::KeyCode::F($code),
+        $crate::crossterm::event::Event::Key($crate::crossterm::event::KeyEvent {
+            code: $crate::crossterm::event::KeyCode::F($code),
             modifiers: $crate::crossterm::modifiers::NONE,
-            kind: ratatui_crossterm::crossterm::event::KeyEventKind::Press|ratatui_crossterm::crossterm::event::KeyEventKind::Repeat,
+            kind: $crate::crossterm::event::KeyEventKind::Press|$crate::crossterm::event::KeyEventKind::Repeat,
             ..
         })
     };
     (keycode press $mod:ident-F($code:literal)) => {
-        ratatui_crossterm::crossterm::event::Event::Key(ratatui_crossterm::crossterm::event::KeyEvent {
-            code: ratatui_crossterm::crossterm::event::KeyCode::F($code),
+        $crate::crossterm::event::Event::Key($crate::crossterm::event::KeyEvent {
+            code: $crate::crossterm::event::KeyCode::F($code),
             modifiers: $crate::crossterm::modifiers::$mod,
-            kind: ratatui_crossterm::crossterm::event::KeyEventKind::Press|ratatui_crossterm::crossterm::event::KeyEventKind::Repeat,
+            kind: $crate::crossterm::event::KeyEventKind::Press|$crate::crossterm::event::KeyEventKind::Repeat,
             ..
         })
     };
     (keycode press $code:ident) => {
-        ratatui_crossterm::crossterm::event::Event::Key(ratatui_crossterm::crossterm::event::KeyEvent {
-            code: ratatui_crossterm::crossterm::event::KeyCode::$code,
+        $crate::crossterm::event::Event::Key($crate::crossterm::event::KeyEvent {
+            code: $crate::crossterm::event::KeyCode::$code,
             modifiers: $crate::crossterm::modifiers::NONE,
-            kind: ratatui_crossterm::crossterm::event::KeyEventKind::Press|ratatui_crossterm::crossterm::event::KeyEventKind::Repeat,
+            kind: $crate::crossterm::event::KeyEventKind::Press|$crate::crossterm::event::KeyEventKind::Repeat,
             ..
         })
     };
     (keycode press $mod:ident-$code:ident) => {
-        ratatui_crossterm::crossterm::event::Event::Key(ratatui_crossterm::crossterm::event::KeyEvent {
-            code: ratatui_crossterm::crossterm::event::KeyCode::$code,
+        $crate::crossterm::event::Event::Key($crate::crossterm::event::KeyEvent {
+            code: $crate::crossterm::event::KeyCode::$code,
             modifiers: $crate::crossterm::modifiers::$mod,
-            kind: ratatui_crossterm::crossterm::event::KeyEventKind::Press|ratatui_crossterm::crossterm::event::KeyEventKind::Repeat,
+            kind: $crate::crossterm::event::KeyEventKind::Press|$crate::crossterm::event::KeyEventKind::Repeat,
             ..
         })
     };
     (keycode release F($code:literal)) => {
-        ratatui_crossterm::crossterm::event::Event::Key(ratatui_crossterm::crossterm::event::KeyEvent {
-            code: ratatui_crossterm::crossterm::event::KeyCode::F($code),
+        $crate::crossterm::event::Event::Key($crate::crossterm::event::KeyEvent {
+            code: $crate::crossterm::event::KeyCode::F($code),
             modifiers: $crate::crossterm::modifiers::NONE,
-            kind: ratatui_crossterm::crossterm::event::KeyEventKind::Release,
+            kind: $crate::crossterm::event::KeyEventKind::Release,
             ..
         })
     };
     (keycode release $mod:ident-F($code:literal)) => {
-        ratatui_crossterm::crossterm::event::Event::Key(ratatui_crossterm::crossterm::event::KeyEvent {
-            code: ratatui_crossterm::crossterm::event::KeyCode::F($code),
+        $crate::crossterm::event::Event::Key($crate::crossterm::event::KeyEvent {
+            code: $crate::crossterm::event::KeyCode::F($code),
             modifiers: $crate::crossterm::modifiers::$mod,
-            kind: ratatui_crossterm::crossterm::event::KeyEventKind::Release,
+            kind: $crate::crossterm::event::KeyEventKind::Release,
             ..
         })
     };
     (keycode release $code:ident) => {
-        ratatui_crossterm::crossterm::event::Event::Key(ratatui_crossterm::crossterm::event::KeyEvent {
-            code: ratatui_crossterm::crossterm::event::KeyCode::$code,
+        $crate::crossterm::event::Event::Key($crate::crossterm::event::KeyEvent {
+            code: $crate::crossterm::event::KeyCode::$code,
             modifiers: $crate::crossterm::modifiers::NONE,
-            kind: ratatui_crossterm::crossterm::event::KeyEventKind::Release,
+            kind: $crate::crossterm::event::KeyEventKind::Release,
             ..
         })
     };
     (keycode release $mod:ident-$code:ident) => {
-        ratatui_crossterm::crossterm::event::Event::Key(ratatui_crossterm::crossterm::event::KeyEvent {
-            code: ratatui_crossterm::crossterm::event::KeyCode::$code,
+        $crate::crossterm::event::Event::Key($crate::crossterm::event::KeyEvent {
+            code: $crate::crossterm::event::KeyCode::$code,
             modifiers: $crate::crossterm::modifiers::$mod,
-            kind: ratatui_crossterm::crossterm::event::KeyEventKind::Release,
+            kind: $crate::crossterm::event::KeyEventKind::Release,
             ..
         })
     };
 
     (mouse down $button:ident for $col:ident, $row:ident ) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::Down(ratatui_crossterm::crossterm::event::MouseButton::$button),
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::Down($crate::crossterm::event::MouseButton::$button),
             column: $col,
             row: $row,
             modifiers: $crate::crossterm::modifiers::NONE,
         })
     };
     (mouse down $mod:ident-$button:ident for $col:ident, $row:ident ) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::Down(ratatui_crossterm::crossterm::event::MouseButton::$button),
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::Down($crate::crossterm::event::MouseButton::$button),
             column: $col,
             row: $row,
             modifiers: $crate::crossterm::modifiers::$mod,
         })
     };
     (mouse up $button:ident for $col:ident, $row:ident ) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::Up(ratatui_crossterm::crossterm::event::MouseButton::$button),
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::Up($crate::crossterm::event::MouseButton::$button),
             column: $col,
             row: $row,
             modifiers: $crate::crossterm::modifiers::NONE,
         })
     };
     (mouse up $mod:ident-$button:ident for $col:ident, $row:ident ) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::Up(ratatui_crossterm::crossterm::event::MouseButton::$button),
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::Up($crate::crossterm::event::MouseButton::$button),
             column: $col,
             row: $row,
             modifiers: $crate::crossterm::modifiers::$mod,
         })
     };
     (mouse drag $button:ident for $col:ident, $row:ident ) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::Drag(ratatui_crossterm::crossterm::event::MouseButton::$button),
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::Drag($crate::crossterm::event::MouseButton::$button),
             column: $col,
             row: $row,
             modifiers: $crate::crossterm::modifiers::NONE,
         })
     };
     (mouse drag $mod:ident-$button:ident for $col:ident, $row:ident ) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::Drag(ratatui_crossterm::crossterm::event::MouseButton::$button),
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::Drag($crate::crossterm::event::MouseButton::$button),
             column: $col,
             row: $row,
             modifiers: $crate::crossterm::modifiers::$mod,
@@ -247,25 +258,25 @@ macro_rules! ct_event {
 
 
     (mouse any for $mouse:ident) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse($mouse)
+        $crate::crossterm::event::Event::Mouse($mouse)
     };
     (mouse any $mod:ident for $mouse:ident ) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse($mouse @ ratatui_crossterm::crossterm::event::MouseEvent {
+        $crate::crossterm::event::Event::Mouse($mouse @ $crate::crossterm::event::MouseEvent {
             modifiers: $crate::crossterm::modifiers::$mod,
             ..
         })
     };
 
     (mouse moved ) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::Moved,
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::Moved,
             modifiers: $crate::crossterm::modifiers::NONE,
             ..
         })
     };
     (mouse moved for $col:ident, $row:ident) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::Moved,
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::Moved,
             column: $col,
             row: $row,
             modifiers: $crate::crossterm::modifiers::NONE,
@@ -273,47 +284,47 @@ macro_rules! ct_event {
     };
 
     (scroll $mod:ident down for $col:ident, $row:ident) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::ScrollDown,
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::ScrollDown,
             column: $col,
             row: $row,
             modifiers: $crate::crossterm::modifiers::$mod,
         })
     };
     (scroll down for $col:ident, $row:ident) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::ScrollDown,
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::ScrollDown,
             column: $col,
             row: $row,
             modifiers: $crate::crossterm::modifiers::NONE,
         })
     };
     (scroll down) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::ScrollDown,
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::ScrollDown,
             modifiers: $crate::crossterm::modifiers::NONE,
             ..
         })
     };
     (scroll $mod:ident up for $col:ident, $row:ident) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::ScrollUp,
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::ScrollUp,
             column: $col,
             row: $row,
             modifiers: $crate::crossterm::modifiers::$mod,
         })
     };
     (scroll up for $col:ident, $row:ident) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::ScrollUp,
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::ScrollUp,
             column: $col,
             row: $row,
             modifiers: $crate::crossterm::modifiers::NONE,
         })
     };
     (scroll up) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::ScrollUp,
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::ScrollUp,
             modifiers: $crate::crossterm::modifiers::NONE,
             ..
         })
@@ -321,50 +332,50 @@ macro_rules! ct_event {
 
     //??
     (scroll left for $col:ident, $row:ident) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::ScrollLeft,
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::ScrollLeft,
             column: $col,
             row: $row,
             modifiers: $crate::crossterm::modifiers::NONE,
         })
     };
     (scroll left) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::ScrollLeft,
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::ScrollLeft,
             modifiers: $crate::crossterm::modifiers::NONE,
             ..
         })
     };
     //??
     (scroll right for $col:ident, $row:ident) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::ScrollRight,
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::ScrollRight,
             column: $col,
             row: $row,
             modifiers: $crate::crossterm::modifiers::NONE,
         })
     };
     (scroll right) => {
-        ratatui_crossterm::crossterm::event::Event::Mouse(ratatui_crossterm::crossterm::event::MouseEvent {
-            kind: ratatui_crossterm::crossterm::event::MouseEventKind::ScrollRight,
+        $crate::crossterm::event::Event::Mouse($crate::crossterm::event::MouseEvent {
+            kind: $crate::crossterm::event::MouseEventKind::ScrollRight,
             modifiers: $crate::crossterm::modifiers::NONE,
             ..
         })
     };
 
     (resized) => {
-        ratatui_crossterm::crossterm::event::Event::Resize(_, _)
+        $crate::crossterm::event::Event::Resize(_, _)
     };
     (resized for $cols:ident, $rows:ident ) => {
-        ratatui_crossterm::crossterm::event::Event::Resize($cols, $rows)
+        $crate::crossterm::event::Event::Resize($cols, $rows)
     };
     (focus_gained) => {
-        ratatui_crossterm::crossterm::event::Event::FocusGained
+        $crate::crossterm::event::Event::FocusGained
     };
     (focus_lost) => {
-        ratatui_crossterm::crossterm::event::Event::FocusLost
+        $crate::crossterm::event::Event::FocusLost
     };
     (paste $value:ident) => {
-        ratatui_crossterm::crossterm::event::Event::Paste($value)
+        $crate::crossterm::event::Event::Paste($value)
     };
 }